fix(studio,runtime,do): close the thermo findings on the assistant seam
Two confirmed defects, one overclaim, and the structural debt the seam added.

**Entry points bypassed the availability latch.** ErrorAlert, every advisor row
and the command palette gated on "is a provider mounted", which is a different
question from "can this deployment run the assistant". On an app with no `AI`
binding all three painted the button, and clicking one fired a real chat RPC
before the panel vanished — the exact failure `aiAvailable` exists to prevent.
The provider now asks once for the shell and every entry point gates on it.

**A seeded question was silently dropped.** `pending` is per-hook, not
per-session, so one turn in flight blocked every send; the seed effect cleared
the ask before `send` could bail, and the operator landed on a blank session with
nothing sent and no error. The ask is now held until the model is free.

**`LUNORA_AI_OPT_IN=disabled` did not disable anything but chat.** The level
reached only the worker-served chat op, so Draft SQL, the filter bar, chart
inference and issue explanation still called the provider, and `aiAvailable`
answered "available" regardless — so every affordance painted. All five now read
one `aiBinding()` helper that returns undefined when the assistant is off.

**Tool results were cut mid-object.** The 20-line log slice was a per-tool fix,
in the transport layer, for a defect every tool has: `runSql` returning forty
rows cut identically, and twenty log entries measured 2773 chars against a 2000
cap, so the stated failure still happened. `fitToBudget` in the engine now drops
whole elements until the payload fits and says how many it omitted.

**`shared/sql-assistant.ts` crossed 1000 lines**, so it splits on the seam its
consumers already have: `shared/ai-prompt.ts` (the inference primitive, the
retry policy, the untrusted boundary and the caps — both halves),
`shared/sql-assistant.ts` (the one-shot SQL tasks — `@lunora/do`), and
`shared/ai-chat.ts` (the transcript budget, the tool catalog, the opt-in ladder
and `generateChat` — `@lunora/runtime`). No consumer imported from both halves.

Also: a blanket rename in the previous commit dragged `SqlAssistantBar` along
with the type it was named after — restored, and the RPC hook is now
`useAssistantRpc`/`AssistantRpc` so it cannot be confused with `useAssistant`,
which returns unrelated shell state. `admin.ts` re-exports four assistant types
instead of hand-copying them (that copy is why `ai-disabled` needed two edits);
the fifth genuinely diverges and says so. `planCell` deleted for the existing
`formatCell`. `AdvisorView` no longer emits an empty spacer span. `newChat` no
longer promises a return value React does not guarantee. Session titles are
translated. `WorkerOptions.aiOptInLevel` is the closed union rather than
`string`. Unused `renameChat`, `toolsFor` and `DEFAULT_AI_OPT_IN_LEVEL` dropped.

Tests: the two confirmed defects each get a case verified to fail without its
fix; plus the `readLogs` tier boundary, the budget fitter, five provider
transitions, and the fenced-SQL parser (moved to `lib/sql-blocks.ts`). The
existing gate-refusal test keeps its top-tier pin.

Plan 364 §10's "no sanitiser needed" claim is narrowed: raw tool output never
replays, but a reply that QUOTES rows does re-enter the client-held transcript.
