+++ api-snapshots/runtime.api.md
@@ -2176,7 +2176,7 @@ interface WorkerOptions {
     adminGate?: (request: Request, context?: ExecutionContextLike) => boolean | Promise<boolean>;
     adminToken?: string;
     aiChatBinding?: AiRunBinding;
-    aiOptInLevel?: string;
+    aiOptInLevel?: AiOptInLevel;
     allowUnauthenticatedShardAccess?: boolean;
     applyGlobals?: GlobalCdcApplyFunction;
     authAdmin?: AuthAdmin;
