+++ packages/studio/src/features/data/data-browser.tsx
@@ -4,7 +4,7 @@ import { useCallback } from "react";
 
 import { ShardInput } from "../../components/shard-input";
 import { useAdminQuery } from "../../hooks/use-admin-query";
-import { useAssistantOps } from "../../hooks/use-assistant-ops";
+import { useAssistantRpc } from "../../hooks/use-assistant-rpc";
 import type { ColumnMeta, FilterClause, TableInfo, TablesColumnsResult } from "../../lib/admin";
 import { ADMIN_FUNCTIONS } from "../../lib/admin";
 import { usePersistedValue } from "../../lib/browser-storage";
@@ -240,7 +240,7 @@ export const DataBrowser = ({
     // Natural-language filtering. The model returns STRUCTURED clauses which
     // land in the visible filter rows for the operator to see and edit — the
     // query never runs off un-reviewed model output.
-    const assistant = useAssistantOps(shardKey);
+    const assistant = useAssistantRpc(shardKey);
 
     const askAiFilter = (prompt: string): void => {
         const apply = async (): Promise<void> => {
