+++ packages/studio/src/features/sql/sql-editor-pane.tsx
@@ -1,9 +1,9 @@
 import type { CSSProperties, ReactElement, RefObject } from "react";
 
-import type { AssistantOps } from "../../hooks/use-assistant-ops";
+import type { AssistantRpc } from "../../hooks/use-assistant-rpc";
 import { useT } from "../../i18n/i18n-context";
 import { EDITOR_TEXT_CLASS } from "./editor-spans";
-import AssistantOpsBar from "./sql-assistant-bar";
+import SqlAssistantBar from "./sql-assistant-bar";
 import type { SqlAutocomplete } from "./sql-autocomplete-ui";
 import { AutocompletePopover } from "./sql-autocomplete-ui";
 import type { SqlDiagnostic } from "./sql-diagnostics";
@@ -22,7 +22,7 @@ const GUTTER_STYLE: CSSProperties = { minWidth: "2.75rem", paddingInline: "0.5re
  * for the reason noted on the props below.
  */
 const SqlEditorPane = ({
-    assistant,
+    rpc,
     autocomplete,
     diagnostics,
     draft,
@@ -36,11 +36,9 @@ const SqlEditorPane = ({
     onRevealDiagnostic,
     overlayRef,
 }: {
-    readonly assistant: AssistantOps;
     /** The completion popover's state, or `null` when closed. */
     readonly autocomplete: SqlAutocomplete["state"];
     readonly diagnostics: ReadonlyArray<SqlDiagnostic>;
-
     readonly draft: string;
 
     /**
@@ -48,6 +46,7 @@ const SqlEditorPane = ({
      * bails out of the whole component when a ref reaches the JSX through a member access.
      */
     readonly editorRef: RefObject<HTMLTextAreaElement | null>;
+
     /** The active tab's last failed statement, which arms the assistant's "Fix this". */
     readonly failed?: { error: string; sql: string };
     readonly gutterRef: RefObject<HTMLDivElement | null>;
@@ -64,13 +63,14 @@ const SqlEditorPane = ({
     readonly onPickSuggestion: (index: number) => void;
     readonly onRevealDiagnostic: (diagnostic: SqlDiagnostic) => void;
     readonly overlayRef: RefObject<HTMLDivElement | null>;
+    readonly rpc: AssistantRpc;
 }): ReactElement => {
     const t = useT();
     const lineCount = draft.split("\n").length;
 
     return (
         <div className="flex min-h-0 min-w-0 flex-1 flex-col">
-            <AssistantOpsBar assistant={assistant} failed={failed} onGenerated={onGenerated} />
+            <SqlAssistantBar failed={failed} onGenerated={onGenerated} rpc={rpc} />
             <div className="flex min-h-0 min-w-0 flex-1">
                 <div
                     aria-hidden="true"
