+++ packages/studio/src/hooks/use-assistant-rpc.ts
@@ -29,7 +29,7 @@ const AI_CHAT = adminRef(ADMIN_FUNCTIONS.aiChat);
 type AssistantTaskKey = "chart" | "chat" | "filter" | "sql";
 
 /** What a surface needs to render one assistant affordance. */
-interface AssistantOps {
+interface AssistantRpc {
     /**
      * One conversational turn. `transcript` is the prior turns, client-held and
      * re-sent; the server caps and fences it, so a long conversation degrades by
@@ -79,7 +79,7 @@ interface AssistantOps {
  * because they are different facts about the deployment, even though the studio
  * does the same thing with both.
  */
-const useAssistantOps = (shardKey: string): AssistantOps => {
+const useAssistantRpc = (shardKey: string): AssistantRpc => {
     const client = useLunora();
 
     const [pendingByTask, setPendingByTask] = useState<Partial<Record<AssistantTaskKey, boolean>>>({});
@@ -137,7 +137,7 @@ const useAssistantOps = (shardKey: string): AssistantOps => {
         }
     };
 
-    const chat: AssistantOps["chat"] = async (prompt, transcript, schema) => {
+    const chat: AssistantRpc["chat"] = async (prompt, transcript, schema) => {
         begin("chat");
 
         try {
@@ -203,5 +203,5 @@ const useAssistantOps = (shardKey: string): AssistantOps => {
     return { chat, generate, inferChart, pending, reason, suggestFilter, unavailable };
 };
 
-export { useAssistantOps };
-export type { AssistantOps, AssistantTaskKey };
+export { useAssistantRpc };
+export type { AssistantRpc, AssistantTaskKey };
