+++ packages/studio/src/lib/admin.ts
@@ -23,6 +23,18 @@ import type { CapturedMail } from "@lunora/mail";
 // the `@lunora/mail` type import above.
 import type { PushSubscriptionDevice } from "@lunora/notify";
 
+/*
+ * The chat transcript and result types come from the ENGINE rather than being
+ * restated here.
+ *
+ * They were duplicated, and had already drifted: the local copy omitted
+ * `toolCalls` and `partial`, so the studio silently dropped the two fields that
+ * say what a turn actually did. A type-only import from `shared/` costs nothing
+ * at runtime and makes that class of drift impossible — the studio already reads
+ * other types from there the same way.
+ */
+import type { GenerateSqlDegradedReason } from "../../../../shared/ai-prompt";
+
 export const ADMIN_FUNCTION_PREFIX = "__lunora_admin__:";
 
 /**
@@ -168,17 +180,9 @@ export interface BulkDeleteResult {
     hasMore: boolean;
 }
 
-/*
- * The chat transcript and result types come from the ENGINE rather than being
- * restated here.
- *
- * They were duplicated, and had already drifted: the local copy omitted
- * `toolCalls` and `partial`, so the studio silently dropped the two fields that
- * say what a turn actually did. A type-only import from `shared/` costs nothing
- * at runtime and makes that class of drift impossible — the studio already reads
- * other types from there the same way.
- */
-export type { ChatResult, ChatTurn, SchemaFact } from "../../../../shared/sql-assistant";
+export type { ChatResult, ChatTurn } from "../../../../shared/ai-chat";
+export type { GenerateSqlDegradedReason, SchemaFact } from "../../../../shared/ai-prompt";
+export type { AssistantChartConfig, GenerateChartResult, GenerateSqlResult } from "../../../../shared/sql-assistant";
 
 /** Result of `__lunora_admin__:getCapturedMail` — the dev mail-catcher inbox, newest first. */
 export interface CapturedMailResult {
@@ -1403,21 +1407,16 @@ export interface SqlConsoleResult {
 }
 
 /**
- * Payload of an `aiGenerateSql` admin call, mirroring `@lunora/do`'s
- * `GenerateSqlResult`. A discriminated union on `degraded` so the ok arm's `sql`
- * is guaranteed by the type rather than by convention.
+ * Payload of an `aiTableFilter` call — structured clauses, never SQL.
+ *
+ * Declared here rather than re-exported with its four siblings above because it
+ * genuinely diverges: the shared arm types `operator` as `string`, and this
+ * narrows it to the {@link FilterOperator} union the browser's filter builder
+ * accepts. The server validates against exactly that set (`FILTER_OPERATORS` in
+ * `shared/sql-assistant.ts`), so the narrowing states a guarantee the server
+ * already enforces and the shared type does not carry.
  */
-export type GenerateSqlResult = { degraded: false; sql: string } | { degraded: true; reason: GenerateSqlDegradedReason };
-
-/** Payload of an `aiTableFilter` call — structured clauses, never SQL. */
 export type GenerateFilterResult = { clauses: FilterClause[]; degraded: false } | { degraded: true; reason: GenerateSqlDegradedReason };
-
-/** A chart the editor can render, inferred from a result set's SHAPE only. */
-export interface AssistantChartConfig {
-    kind: "area" | "bar" | "line";
-    x: string;
-    y: string[];
-}
 
 /** Whether the deployment has an `AI` binding — the gate for every assistant affordance. */
 export interface AiAvailableResult {
@@ -1434,12 +1433,6 @@ export interface AiAvailableResult {
  * so a field added in `@lunora/notify` flows here automatically.
  */
 export type { PushSubscriptionDevice } from "@lunora/notify";
-
-/** Payload of an `aiChartConfig` call. */
-export type GenerateChartResult = { chart: AssistantChartConfig; degraded: false } | { degraded: true; reason: GenerateSqlDegradedReason };
-
-/** Why no statement came back. `no-ai-binding` means the app has no AI binding — hide the affordance entirely. */
-export type GenerateSqlDegradedReason = "ai-disabled" | "ai-error" | "empty-response" | "no-ai-binding" | "unsafe-response";
 
 /** One reverse edge: `table.column` holds a foreign key pointing at the browsed table. */
 export interface BackRelation {
