+++ packages/studio/src/locales/en.ts
@@ -37,12 +37,17 @@ const MESSAGE_IDS = [
     "Assistant",
     "Close assistant",
     "Close chat",
+    "Debug error",
+    "Debug query",
     "Debug with AI",
+    "Explain query",
     "Explain this query",
     "Insert into editor",
     "New chat",
     "Older turns were dropped to fit the context budget.",
+    "Read plan",
     "Read this plan",
+    "SQL console",
     "Stopped early — this answer is incomplete.",
     "The model could not be reached.",
     "The model returned a statement that is not read-only, so it was discarded.",
